Roles para acciones, sweetalert

+++ GestionArchivo/src/components/TaskCard.jsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import { useTasks } from "../context/TasksContext";
+import { useAuth } from "../context/AuthContext";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import {
   faTrashAlt,
@@ -11,16 +12,28 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 import "./Table.css";
 import Paginator from "./Paginator";
+import Swal from 'sweetalert2';
 
 function Table() {
   const { tasks, deleteTask } = useTasks();
+  const { user } = useAuth(); // Suponiendo que 'user' contiene el rol del usuario
   const [search, setSearch] = useState("");
   const [currentPage, setCurrentPage] = useState(1);
   const tasksPerPage = 10;
 
-  const handleDelete = (id) => {
-    deleteTask(id);
-  };
+  async function handleDelete(id) {
+    Swal.fire({
+      title: '¿Está seguro que desea eliminar este archivo?',
+      confirmButtonText: 'Eliminar',
+      denyButtonText: 'Cancelar',
+      showDenyButton: true,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await deleteTask(id); // Asegúrate de que `deleteTask` es la función correcta
+        Swal.fire({ title: "Éxito", text: "Archivo eliminado correctamente", icon: "success", timer: 3500 });
+      }
+    });
+  }
 
   const searcher = (e) => {
     setSearch(e.target.value);
@@ -30,9 +43,7 @@ function Table() {
     ? tasks
     : tasks.filter((task) => {
         const expe = task.expe ? task.expe.toLowerCase() : "";
-        const correlativo = task.correlativo
-          ? task.correlativo.toLowerCase()
-          : "";
+        const correlativo = task.correlativo ? task.correlativo.toLowerCase() : "";
         const anio = task.anio ? task.anio.toLowerCase() : "";
         const cuerpo = task.cuerpo ? task.cuerpo.toLowerCase() : "";
         const fecha = task.fecha ? task.fecha.toLowerCase() : "";
@@ -51,7 +62,6 @@ function Table() {
         );
       });
 
-  // Paginación
   const indexOfLastTask = currentPage * tasksPerPage;
   const indexOfFirstTask = indexOfLastTask - tasksPerPage;
   const currentTasks = filteredTasks.slice(indexOfFirstTask, indexOfLastTask);
@@ -62,24 +72,21 @@ function Table() {
 
   const formatDate = (dateStr) => {
     if (!dateStr) return "";
-    
-    // Crear un objeto Date en UTC
+
     const utcDate = new Date(dateStr);
-    
-    // Obtener el desfase horario de la zona de Buenos Aires en minutos
-    const offset = 3 * 60; // -3 horas en minutos
-    
-    // Crear una nueva fecha en la zona horaria local
+    const offset = 3 * 60;
     const localDate = new Date(utcDate.getTime() + offset * 60 * 1000);
-    
-    // Obtener el día, mes y año en la zona horaria local
     const day = String(localDate.getUTCDate()).padStart(2, '0');
-    const month = String(localDate.getUTCMonth() + 1).padStart(2, '0'); // Meses empiezan desde 0
+    const month = String(localDate.getUTCMonth() + 1).padStart(2, '0');
     const year = localDate.getUTCFullYear();
-    
-    // Formatear la fecha en 'dd/MM/yyyy'
+
     return `${day}/${month}/${year}`;
   };
+
+  const canView = ['admin', 'user', 'boss', 'viewer'].includes(user.role);
+  const canEdit = ['admin', 'user'].includes(user.role);
+  const canDelete = user.role === 'admin';
+  const canAddTask = !['boss', 'viewer'].includes(user.role);
 
   return (
     <div className="container-fluid bg-primary vh-100 vw-100 d-flex align-items-center justify-content-center">
@@ -96,20 +103,19 @@ function Table() {
                 className="form-control border border-gray-300 focus:outline-none"
                 aria-describedby="search-icon"
               />
-              <span
-                className="input-group-text bg-white border-none"
-                id="search-icon"
-              >
+              <span className="input-group-text bg-white border-none" id="search-icon">
                 <FontAwesomeIcon icon={faSearch} className="ml-2" />
               </span>
             </div>
-            <ul className="button-container">
-              <li>
-                <Link to="/add-task" className="btn btn-success">
-                  <FontAwesomeIcon icon={faPlus} />
-                </Link>
-              </li>
-            </ul>
+            {canAddTask && (
+              <ul className="button-container">
+                <li>
+                  <Link to="/add-task" className="btn btn-success">
+                    <FontAwesomeIcon icon={faPlus} />
+                  </Link>
+                </li>
+              </ul>
+            )}
             <br />
             <div className="table-scroll">
               <table className="table" style={{ textTransform: "uppercase" }}>
@@ -128,57 +134,45 @@ function Table() {
                     <th>Iniciador</th>
                     <th>Asunto</th>
                     <th>Ver</th>
-                    <th>Editar</th>
-                    <th>Borrar</th>
+                    {canEdit && <th>Editar</th>}
+                    {canDelete && <th>Borrar</th>}
                   </tr>
                 </thead>
                 <tbody>
                   {currentTasks.map((task) => (
                     <tr key={task._id}>
-                      <td data-label ="Codigo de Organismo">{task.expe ? task.expe.toUpperCase() : ""}</td>
-                      <td data-label ="N° Correlativo">
-                        {task.correlativo ? task.correlativo.toUpperCase() : ""}
-                      </td>
+                      <td data-label="Codigo de Organismo">{task.expe ? task.expe.toUpperCase() : ""}</td>
+                      <td data-label="N° Correlativo">{task.correlativo ? task.correlativo.toUpperCase() : ""}</td>
                       <td data-label="Año">{task.anio ? task.anio.toUpperCase() : ""}</td>
                       <td data-label="Cuerpo">{task.cuerpo ? task.cuerpo.toUpperCase() : ""}</td>
-                      <td data-label="Fecha">
-                        {formatDate(task.fecha)}
-                      </td>
-                      <td data-label="Iniciador">
-                        {task.iniciador ? task.iniciador.toUpperCase() : ""}
-                      </td>
+                      <td data-label="Fecha">{formatDate(task.fecha)}</td>
+                      <td data-label="Iniciador">{task.iniciador ? task.iniciador.toUpperCase() : ""}</td>
                       <td data-label="Asunto">{task.asunto ? task.asunto.toUpperCase() : ""}</td>
-
                       <td data-label="Ver">
                         <div className="button-container">
-                          <Link
-                            className="btn btn-success"
-                            to={`/view/task/${task._id}`}
-                          >
+                          <Link className="btn btn-success" to={`/view/task/${task._id}`}>
                             <FontAwesomeIcon icon={faEye} />
                           </Link>
                         </div>
                       </td>
-                      <td data-label="Editar">
-                        <div className="button-container">
-                          <Link
-                            className="btn btn-primary"
-                            to={`/edit-task/${task._id}`}
-                          >
-                            <FontAwesomeIcon icon={faEdit} />
-                          </Link>
-                        </div>
-                      </td>
-                      <td data-label="Borrar">
-                        <div className="button-container">
-                          <button
-                            className="btn btn-danger"
-                            onClick={() => handleDelete(task._id)}
-                          >
-                            <FontAwesomeIcon icon={faTrashAlt} />
-                          </button>
-                        </div>
-                      </td>
+                      {canEdit && (
+                        <td data-label="Editar">
+                          <div className="button-container">
+                            <Link className="btn btn-primary" to={`/edit-task/${task._id}`}>
+                              <FontAwesomeIcon icon={faEdit} />
+                            </Link>
+                          </div>
+                        </td>
+                      )}
+                      {canDelete && (
+                        <td data-label="Borrar">
+                          <div className="button-container">
+                            <button className="btn btn-danger" onClick={() => handleDelete(task._id)}>
+                              <FontAwesomeIcon icon={faTrashAlt} />
+                            </button>
+                          </div>
+                        </td>
+                      )}
                     </tr>
                   ))}
                 </tbody>
@@ -197,3 +191,4 @@ function Table() {
 }
 
 export default Table;
+
